Extract utility functions to separate file
